==> 00_execution_hub.ipynb <==
import nbformat
from nbconvert.preprocessors import ExecutePreprocessor as EP

import os
os.listdir()

notebooks = ['01_static_data_preparation.ipynb',
             '02_infection_data_preparation.ipynb',
#              '03_mobility_data_preparation.ipynb', #these don't work without access
#              '04a_mobility_feature_creation.ipynb', #to proprietary mobility data
             '04b_weather_feature_creation.ipynb',
             '05_target_and_prediction.ipynb',
             '06_deviance_analysis.ipynb',
             '08_visualization.ipynb']

ep_ = EP(timeout=20000)

for nb_file in notebooks:
    print('Now working on', nb_file)
    with open(nb_file) as f:
        nb = nbformat.read(f, nbformat.NO_CONVERT)
        try:
            ep_.preprocess(nb)
        except:
            print('Execuction Error in', nb_file)
