==> 03_mobility_data_preparation.ipynb <==
# Script for transforming raw mobility data to edge table and trips per day and district

## Specify identifier for output files

identifier = 'teralytics_2019_20_'

## Import libraries

import pandas as pd
import numpy as np

## Read in raw data from file

Read in file into dataframe:

df_raw = pd.DataFrame()

for chunk in pd.read_csv('data/deldd-1-taeglich-aktualisierte-langdistanz-reisen-giz-counts.csv', chunksize=10**6):
    df_raw = df_raw.append(chunk)

df_raw.reset_index(inplace=True, drop=True)

df_raw.head(5)

Drop columns we don't need:

df_prep = df_raw.drop(columns=['Mode of Transport','Trip distance (km)', 'StartName', 'EndName'])

Specify columns names with origin and destination districts, dates and mobility counts:

origin_col_name = 'StartId'
destination_col_name = 'EndId'
count_col_name = 'Count'


## OSM ID to AGS5

o2a_df = pd.read_csv('OSM_2_AGS_v2.csv')

osm_2_ags_dict = dict(zip(o2a_df.OSM_ID,o2a_df.AGS_5))

df_prep['orig_ags5'] = df_prep[origin_col_name].map(osm_2_ags_dict)
df_prep['dest_ags5'] = df_prep[destination_col_name].map(osm_2_ags_dict)

import locale
locale.setlocale(category=locale.LC_ALL,locale="German")

df_prep['date'] = pd.to_datetime(df_prep['Bucket'], format = '%Y-%m-%d (%A)')

df_prep.drop(columns=[origin_col_name,destination_col_name, "Bucket"], inplace=True)

df_prep.head(5)

## Create edge table

edge_table = df_prep.groupby(['date','orig_ags5','dest_ags5'])[count_col_name].sum().reset_index()

Save edge table to CSV-file using the specifier defined above:

file_name = identifier + 'edge_table.csv'
edge_table.to_csv(file_name, index=False)

In case there's a problem later on, reading in this CSV will serve as a checkpoint:

#edge_table = pd.read_csv(file_name)
#edge_table.tail(10)

## Create table for trips per district and day

Create Boolean Series to subset data into within- and between-district mobility:

is_internal = edge_table.orig_ags5 == edge_table.dest_ags5

Create multi-indexed Series for outgoing, incoming and internal mobility per district and day:

outgoing_mobility = edge_table[~is_internal].groupby(['date', 'orig_ags5'])[count_col_name].sum().reset_index()
incoming_mobility = edge_table[~is_internal].groupby(['date', 'dest_ags5'])[count_col_name].sum().reset_index()
internal_mobility = edge_table[ is_internal].groupby(['date', 'orig_ags5'])[count_col_name].sum().reset_index()

Merge all three into one dataframe, clean that dataframe up, and add total mobility:

external_mobility = pd.merge(outgoing_mobility, incoming_mobility, 
                   left_on = ['date', 'orig_ags5'], right_on = ['date', 'dest_ags5'],
                   suffixes = ('_outgoing','_incoming'))

mobility_per_district_day = pd.merge(external_mobility, internal_mobility,
                                    on = ['date', 'orig_ags5'])

#rename columns:
mobility_per_district_day.rename(columns = {count_col_name : count_col_name + '_internal',
                                           'orig_ags5' : 'districtId'}, 
                                 inplace=True)

#drop redundant dest_ags5 column:
mobility_per_district_day.drop(columns = 'dest_ags5', inplace=True)

#add total_mobility as sum:
count_cols = [count_col_name + suffix for suffix in ['_outgoing','_incoming','_internal']]
mobility_per_district_day['total_mobility'] = mobility_per_district_day[count_cols].sum(axis=1)

Save mobility per district and day to CSV, again using the identifier:

file_name = identifier + 'mobility_per_district_and_day.csv'
mobility_per_district_day.to_csv(file_name, index=False)
