==> 09_data_tables.ipynb <==
import numpy as np
import pandas as pd

all_data = pd.read_csv('all_data.csv')

## descriptives

cols_to_include = ['target',
                   'R0_this_week',
                   'Count_internal_p_pop',
                   'Count_incoming_p_pop',
                   'incoming_infected_p_pop',
                   'temperature',
                   'humidity',
                   'precipitation',
                   'sunshine',
                   'ses_z',
                   'rur_z']

descriptions = (all_data[cols_to_include]
                .describe()
                .T
                [['mean', 'std', 'min', '25%', '50%', '75%', 'max']])

ex=5 # exponent to make infection load visibly non-zero

descriptions.loc['incoming_infected_p_pop'] = descriptions.loc['incoming_infected_p_pop'] * (10**ex)

var_dict = {'target':'Log of R-factor',
            'R0_this_week':'R-factor',
            'Count_internal_p_pop':'Internal trips per inhabitant',
            'Count_incoming_p_pop':'Incoming trips per inhabitant',
            'incoming_infected_p_pop':'Incoming infection load (10^-'+str(ex)+' per inhabitant)',
            'temperature':'Average temperature (°C)',
            'humidity':'Average humidity (%)',
            'precipitation':'Average daily precipitation (cm)',
            'sunshine':'Average daily sunshine (minutes)',
            'ses_z':'Standardized socio-economic status index',
            'rur_z':'Standardized ruralness index'}

descriptions.index=descriptions.index.map(var_dict)
descriptions = descriptions.round(2)
print(descriptions)
descriptions.to_csv('descriptives.csv')

## infection numbers 

cols_to_include = ['districtId',
                   'week_no',
                   'AnzahlFall',
                  'AnzahlTodesfall',
                   'AnzahlFall_per_cap',
                  'AnzahlTodesfall_per_cap',
                   'AnzahlFall_cumul',
                  'AnzahlTodesfall_cumul',
                   'AnzahlFall_cumul_per_cap',
                  'AnzahlTodesfall_cumul_per_cap',
                  'R0_this_week',
                   'target']

infections = all_data[cols_to_include]

var_dict = {'districtId':'ID/AGS-5',
            'week_no':'Calendar week 2020',
            'AnzahlFall':'Number of new cases',
            'AnzahlTodesfall':'Number of new fatalities',
            'AnzahlFall_per_cap':'New cases per inhabitant',
            'AnzahlTodesfall_per_cap':'New fatalities per inhabitant',
            'AnzahlFall_cumul':'Cumulated number of cases',
            'AnzahlTodesfall_cumul':'Cumulated number of fatalities',
            'AnzahlFall_cumul_per_cap':'Cumulated cases per inhabitant',
            'AnzahlTodesfall_cumul_per_cap':'Cumulated fatalities per inhabitant',
            'R0_this_week':'Weekly R0',
            'target':'Weekly Logarithmic R0'}

infections.rename(columns=var_dict).to_csv('infection_numbers.csv',index=False)

## pd descriptives

pd_data = pd.merge(pd.read_csv('deviance_df.csv'),
                   pd.read_csv('processed_static_data.csv'),
                   on='districtId',
                   suffixes=('','_duplicate')).query("is_lockdown_pd or is_summer_pd")
pd_data['BL_code'] = pd_data.districtId // 1000
pd_data['Province'] = pd_data.BL_code.map({3:'Lower Saxony',
                                          5:'North Rine-Westphalia',
                                          7:'Rhineland-Palatinate',
                                          8:'Baden-Wuerttemberg',
                                          9:'Bavaria'})
pd_data['PD Timeframe'] = pd_data.is_lockdown_pd.map({True:'First Lockdown',
                                                False:'Summer Vacation'})
pd_data['District'] = pd_data.district_name.str.replace('KS ','Independent city of ')
pd_data.rename(columns={'lockdown_rank':'Ranking during first lockdown',
                       'summer_rank':'Ranking during summer vacation',
                       'rur_z':'Ruralness',
                       'ses_z':'Socioeconimic status',
                       'total_population':'Population',
                       'population_density':'Population density (inhabitants per km²)'},
              inplace=True)

for c in pd_data.columns:
    try:
        if all(pd_data[c] % 1 == 0):
            pd_data[c] = pd_data[c].astype(int)
        else:
            pd_data[c] = pd_data[c].round(2)
    except:
        pass
    
pd_data

cols_2_desc = ['District',
               'PD Timeframe',
               'Province',
               'Ranking during first lockdown',
               'Ranking during summer vacation',
               'Ruralness',
               'Socioeconimic status',
               'Population',
               'Population density (inhabitants per km²)']
district_order = [142,211,294,98,256,58]

pd_data[cols_2_desc].loc[district_order].to_csv('pd_descriptives.csv',index=False)
